lab3-langchain-agent/pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "LangChain Agent для AI Agents платформы"
authors = [
    {name = "evo-ai-agents-labs", email = "labs@cloud.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "a2a-sdk>=0.3.4,<0.4.0",
    "litellm>=1.76.3,<2.0.0",
    "openinference-instrumentation-langchain>=0.1.0,<0.2.0",
    "arize-phoenix-otel>=0.13.0,<0.14.0",
    "httpx>=0.27.0,<1.0.0",
    "uvicorn>=0.30.0,<1.0.0",
]


